+++ examples/openapi/app/docs/layout.tsx
@@ -5,14 +5,7 @@ import { baseOptions } from '@/app/layout.config';
 
 export default function RootDocsLayout({ children }: { children: ReactNode }) {
   return (
-    <DocsLayout
-      tree={source.pageTree}
-      {...baseOptions}
-      nav={{
-        ...baseOptions.nav,
-        mode: 'top',
-      }}
-    >
+    <DocsLayout tree={source.pageTree} {...baseOptions}>
       {children}
     </DocsLayout>
   );