+++ examples/openapi/scripts/generate-docs.mjs
@@ -1,18 +1,22 @@
 import * as OpenAPI from 'fumadocs-openapi';
-import { rimrafSync } from 'rimraf';
+import { rimraf } from 'rimraf';
 
 const out = './content/docs/(api)';
 
-// clean generated files
-rimrafSync(out, {
-  filter(v) {
-    return !v.endsWith('index.mdx') && !v.endsWith('meta.json');
-  },
-});
+async function generate() {
+  // clean generated files
+  await rimraf(out, {
+    filter(v) {
+      return !v.endsWith('index.mdx') && !v.endsWith('meta.json');
+    },
+  });
 
-void OpenAPI.generateFiles({
-  // input files
-  input: ['./openapi.json'],
-  output: out,
-  groupBy: 'tag',
-});
+  await OpenAPI.generateFiles({
+    // input files
+    input: ['./openapi.json'],
+    output: out,
+    groupBy: 'tag',
+  });
+}
+
+void generate();