+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # next-docs-zeta
 
+## 15.2.8
+
 ## 15.2.7
 
 ### Patch Changes